+++ api/params/network_create.go
@@ -10,4 +10,8 @@ type NetworkCreate struct {
 	Type    types.NetworkType `json:"type"`
 	IPRange string            `json:"ip_range"`
 	Gateway string            `json:"gateway"`
+
+	// Parameter injected by docker integration only
+	// in order to tag the network.
+	CreatedByDocker bool `json:"-"`
 }