+++ web/networks_controller_destroy.go
@@ -39,9 +39,21 @@ func (c NetworksController) Destroy(w http.ResponseWriter, r *http.Request, p ma
 		return errors.Errorf("fail to delete network %s, %d endpoints are still present.", n, len(endpoints))
 	}
 
+	err = c.NetworkRepository.Deactivate(ctx, n)
+	if err != nil {
+		return errors.Wrapf(err, "fail to deactivate network %s", n)
+	}
+
 	err = c.NetworkRepository.Delete(ctx, n)
 	if err != nil {
 		return errors.Wrapf(err, "fail to delete network %s", n)
+	}
+
+	if !n.CreatedByDocker {
+		err = c.IPAllocator.ReleasePool(ctx, n.ID)
+		if err != nil {
+			return errors.Wrapf(err, "fail to release pool of network %s", n)
+		}
 	}
 
 	w.WriteHeader(204)